anymind-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anymind"
version = "0.1.0"
description = "Anymind Python SDK and CLI for deploying AI agents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Anymind", email = "support@anymind.com"}
]
keywords = ["ai", "agents", "deployment", "mlops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "typer>=0.9.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
anymind = "anymind.cli:main"

[project.urls]
Homepage = "https://github.com/anymind/anymind-sdk"
Documentation = "https://docs.anymind.com"
Repository = "https://github.com/anymind/anymind-sdk"
Issues = "https://github.com/anymind/anymind-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["anymind*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

